+++ src/commands/removeFile.py
@@ -1,4 +1,5 @@
 from src.objs import *
+from src.functions.floodControl import floodControl
 from src.functions.exceptions import exceptions, noAccount
 
 #: Remove files
@@ -6,24 +7,26 @@ from src.functions.exceptions import exceptions, noAccount
 def removeFile(message):
     userId = message.from_user.id
     userLanguage = dbSql.getSetting(userId, 'language')
-    ac = dbSql.getDefaultAc(userId)
 
-    #! If user has an account
-    if ac:
-        sent = bot.send_message(message.chat.id, language['removingFile'][userLanguage])
-        id = message.text[8:]
+    if floodControl(message, userLanguage):
+        ac = dbSql.getDefaultAc(userId)
 
-        account = Seedr(cookie=ac['cookie'])
+        #! If user has an account
+        if ac:
+            sent = bot.send_message(message.chat.id, language['removingFile'][userLanguage])
+            id = message.text[8:]
 
-        response = account.deleteFile(id).json()
+            account = Seedr(cookie=ac['cookie'])
 
-        #! If file removed successfully
-        if response['result'] == True:
-            bot.edit_message_text(text=language['removedSuccessfully'][userLanguage], chat_id=message.chat.id, message_id=sent.id)
-        
+            response = account.deleteFile(id).json()
+
+            #! If file removed successfully
+            if response['result'] == True:
+                bot.edit_message_text(text=language['removedSuccessfully'][userLanguage], chat_id=message.chat.id, message_id=sent.id)
+            
+            else:
+                exceptions(message, response, userLanguage)
+            
+        #! If no accounts
         else:
-            exceptions(message, response, userLanguage)
-        
-    #! If no accounts
-    else:
-        noAccount(message, userLanguage)
+            noAccount(message, userLanguage)
